Fix #15664, 568c7bb2: regression in aircraft movement

+++ src/direction_func.h
@@ -71,7 +71,7 @@ inline DirDiff DirDifference(Direction d0, Direction d1)
 	assert(IsValidDirection(d1));
 	/* Cast to uint so compiler can use bitmask. If the difference is negative
 	 * and we used int instead of uint, further "+ 8" would have to be added. */
-	return static_cast<DirDiff>((to_underlying(d0) - to_underlying(d1)) % 8);
+	return static_cast<DirDiff>(static_cast<uint>(to_underlying(d0) - to_underlying(d1)) % 8);
 }
 
 /**
@@ -88,7 +88,7 @@ inline DirDiff DirDifference(Direction d0, Direction d1)
 inline DirDiff ChangeDirDiff(DirDiff d, DirDiff delta)
 {
 	/* Cast to uint so compiler can use bitmask. Result can never be negative. */
-	return static_cast<DirDiff>((to_underlying(d) + to_underlying(delta)) % 8);
+	return static_cast<DirDiff>(static_cast<uint>(to_underlying(d) + to_underlying(delta)) % 8);
 }
 
 /**
@@ -115,7 +115,7 @@ inline Direction ChangeDir(Direction d, DirDiff delta)
 {
 	assert(IsValidDirection(d));
 	/* Cast to uint so compiler can use bitmask. Result can never be negative. */
-	return static_cast<Direction>((to_underlying(d) + to_underlying(delta)) % 8);
+	return static_cast<Direction>(static_cast<uint>(to_underlying(d) + to_underlying(delta)) % 8);
 }
 
 
@@ -143,7 +143,7 @@ inline DiagDirDiff DiagDirDifference(DiagDirection d0, DiagDirection d1)
 	assert(IsValidDiagDirection(d0));
 	assert(IsValidDiagDirection(d1));
 	/* Cast to uint so compiler can use bitmask. Result can never be negative. */
-	return static_cast<DiagDirDiff>((to_underlying(d0) - to_underlying(d1)) % 4);
+	return static_cast<DiagDirDiff>(static_cast<uint>(to_underlying(d0) - to_underlying(d1)) % 4);
 }
 
 /**
@@ -160,7 +160,7 @@ inline DiagDirection ChangeDiagDir(DiagDirection d, DiagDirDiff delta)
 {
 	assert(IsValidDiagDirection(d));
 	/* Cast to uint so compiler can use bitmask. Result can never be negative. */
-	return static_cast<DiagDirection>((to_underlying(d) + to_underlying(delta)) % 4);
+	return static_cast<DiagDirection>(static_cast<uint>(to_underlying(d) + to_underlying(delta)) % 4);
 }
 
 /**
